Move `landing/components` to `landing/sections`
